Signals dialog: make it possible to compose derivations with signal factors that are not an integer.  E.g.: It's now possible to create a signalcomposition of (1 x F4) - (0.5 x A1) - (0.5 x A2).

+++ global.h
@@ -196,7 +196,7 @@ struct signalcompblock{
         double sample_timeoffset_part;
         double sample_pixel_ratio;
         int edfsignal[MAXSIGNALS];
-        int factor[MAXSIGNALS];
+        double factor[MAXSIGNALS];
         int polarity;
         double sensitivity[MAXSIGNALS];
         int oldvalue;
